+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [SQL] CREATE LOCAL VIEW statement to declare views which are not outputs.
+  ([#1851](https://github.com/feldera/feldera/pull/1851))
+
 ### Added
 
 - Documentation: added the